Handling single user Workflow
git-svn-id: http://192.168.110.32/svn/Products/PFF/Base/Branches/MLOD@88540 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/service/customermasters/impl/CustomerDetailsServiceImpl.java
@@ -2652,8 +2652,7 @@ public class CustomerDetailsServiceImpl extends GenericService<Customer> impleme
 
 		List<AuditDetail> auditDetailList = new ArrayList<AuditDetail>();
 
-		if (!StringUtils.equals(customerDetails.getSourceId(), PennantConstants.FINSOURCE_ID_API)) {
-			//getListAuditDetails(listDeletion(customerDetails, "_Temp", auditHeader.getAuditTranType()))
+		if (!StringUtils.equals(customerDetails.getSourceId(), PennantConstants.FINSOURCE_ID_API) && !customerDetails.isNewRecord()) {
 			auditDetailList.addAll(listDeletion(customerDetails, "_Temp", auditHeader.getAuditTranType()));
 			getCustomerDAO().delete(customer, "_Temp");
 			auditHeader.setAuditTranType(PennantConstants.TRAN_WF);
